requirements.txt
asgiref
autopep8
Django>=3.0.7
django-cors-headers>=3.2.1
django-crispy-forms>=1.9.0
django-extensions>=2.2.9
django-filter==2.4.0
# django-filters==0.2.1
djangorestframework>=3.12.1
httplib2>=0.18.1
psycopg2-binary>=2.8.4
pytz>=2020.1
urllib3>=1.25.11
Werkzeug>=1.0.1
beautifulsoup4>=4.9.3
pycodestyle
sqlparse
Unipath>=1.1
dj-database-url
python-decouple
gunicorn
whitenoise
pytest>=6.1.1
pytest-django >= 4.1.0
requests>=2.24.0
lxml>=4.6.1